Updated readme.md -- adjusted plan
Updated plan according to new knowledge:
* core code does NOT need to have robot "enter" goal (Arthur)
* wiki discussion = checking on commits & all.

Thus, this moves the timeline a bit. I also removed the "due date of the objective" since that comes down to when the code and when the report is due. I clarified on tasks; the plan will be that during the 1/2 week after the code for one part works, it'll be tested + documented by two members of the team, while the others focus on the rest code.

Meetings stay the same, so just... do like I tell you?

Anyways, stay safe xx